Apply recent light source interface
